Bump form state up to App.jsx
