+++ app/javascript/components/intersecting-streets/ChildSurveys.js
@@ -1,49 +1,15 @@
-import React, { useReducer } from 'react';
+import React from 'react';
 import ChildSurvey from './ChildSurvey';
 import { Button } from 'semantic-ui-react';
 
-function reducer(state, action) {
-  switch(action.type) {
-    case 'updateStudent':
-      const updatedStudentInfo = state.studentInfo
-      updatedStudentInfo[`${action.id}`][`${action.property}`] = action.value;
-      return { studentInfo: updatedStudentInfo }
-    case 'addStudent':
-      const increasedStudentInfo = state.studentInfo
-      increasedStudentInfo.push({
-        grade: '',
-        to_school: '',
-        dropoff: '',
-        from_school: '',
-        pickup: ''
-      })
-      return { studentInfo: increasedStudentInfo }
-    case 'removeStudent':
-      const removedStudentInfo = state.studentInfo
-      removedStudentInfo.splice([`${action.id}`], 1)
-      return { studentInfo: removedStudentInfo}
-    default:
-      throw new Error();
-  }
-}
-
-const ChildSurveys = () => {
-  const [state, dispatch] = useReducer(reducer, {
-    studentInfo: [{
-      grade: '',
-      to_school: '',
-      dropoff: '',
-      from_school: '',
-      pickup: '',
-    }]
-  })
+const ChildSurveys = ({studentInfo, dispatch}) => {
   let childSurveys = [];
-  for (var i = 0; i < state.studentInfo.length; i++) {
+  for (var i = 0; i < studentInfo.length; i++) {
     childSurveys.push(
       <ChildSurvey
         id={i}
         key={i}
-        studentInfo={state.studentInfo}
+        studentInfo={studentInfo}
         dispatch={dispatch}
       />
     );
